requirements.txt
Flask==2.3.3
Flask-Cors==3.0.10
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==7.4.2


